Improve form model for strands, isolate org questions
